aider: Add the `rpg-toast` class to the `_MyHome.scss` file.
# Aider chat conversation:

USER: add the css class to the stylesheet

ASSISTANT: Sure, we can add the `rpg-toast` class to the `_MyHome.scss` file. This class will style the toast to make it look like an RPG dialog box.

Here is the change:

```scss
src/views/MyPortal/MyHome/_MyHome.scss
<<<<<<< ORIGINAL
.my-home{
  ion-content{
    --background: url("@/assets/images/backgrounds/my-home.jpg") center center/cover no-repeat;
    ion-card{
      ion-button{
        &::part(native){
        }
      }
    }
  }
}
=======
.my-home{
  ion-content{
    --background: url("@/assets/images/backgrounds/my-home.jpg") center center/cover no-repeat;
    ion-card{
      ion-button{
        &::part(native){
        }
      }
    }
  }
}

.rpg-toast {
  --background: rgba(0, 0, 0, 0.8);
  --color: #fff;
  --border-radius: 10px;
  --padding: 10px;
  --width: 90%;
  --max-height: 150px;
  --min-height: 50px;
  --margin-bottom: 20px;
}
>>>>>>> UPDATED
```

This will style the toast with a semi-transparent black background, white text, rounded corners, and some padding. The toast will take up 90% of the screen width, and will be positioned 20px from the bottom of the screen. The height of the toast will be between 50px and 150px, depending on the content.

Please adjust these values as needed to fit your design.