aider: Updated "Who" select to allow multiple selections and disabled it when the type is "asNeeded".
# Aider chat conversation:

USER: i need the option to be able to select multiple users when using choosing who. whenever, the type is changed to "as needed" the who should be disabled as it will be assigned to everyone.

ASSISTANT: To implement this feature, we need to modify the following files:

- src/views/App/SideMenu/XpGameMaster/XpAchievements/XpAddAchievement/XpAddAchievement.vue
- src/views/App/SideMenu/XpGameMaster/XpAchievements/XpAddAchievement/XpAddAchievement.ts

In the XpAddAchievement.vue file, we will change the "Who" `ion-select` to allow multiple selections by adding the `multiple` attribute. We will also add a `:disabled` attribute to disable the "Who" select when the type is "asNeeded".

In the XpAddAchievement.ts file, we will change the `assignee` ref to an array to hold multiple values.

Here are the changes:

```vue
src/views/App/SideMenu/XpGameMaster/XpAchievements/XpAddAchievement/XpAddAchievement.vue
<<<<<<< ORIGINAL
                  <ion-item>
                    <ion-label>Who</ion-label>
                    <ion-select v-model="assignee">
                      <!-- Replace with actual list of users -->
                      <ion-select-option value="user1">User 1</ion-select-option>
                      <ion-select-option value="user2">User 2</ion-select-option>
                    </ion-select>
                  </ion-item>
=======
                  <ion-item>
                    <ion-label>Who</ion-label>
                    <ion-select v-model="assignee" multiple :disabled="type === 'asNeeded'">
                      <!-- Replace with actual list of users -->
                      <ion-select-option value="user1">User 1</ion-select-option>
                      <ion-select-option value="user2">User 2</ion-select-option>
                    </ion-select>
                  </ion-item>
>>>>>>> UPDATED
```

```ts
src/views/App/SideMenu/XpGameMaster/XpAchievements/XpAddAchievement/XpAddAchievement.ts
<<<<<<< ORIGINAL
    const assignee = ref('');
=======
    const assignee = ref([]);
>>>>>>> UPDATED
```

Please confirm these changes.