aider: Added namespace 'com.forthexp.doit' to the android block in capacitor-cordova-android-plugins/build.gradle.
# Aider chat conversation:

ASSISTANT: The error message indicates that the namespace is not specified in the `android` block of the `capacitor-cordova-android-plugins` module's `build.gradle` file.

To fix this issue, you need to add the `namespace` attribute to the `android` block in the `capacitor-cordova-android-plugins/build.gradle` file. The value of the `namespace` attribute should be the same as the `package` attribute in your `AndroidManifest.xml` file.

Here is the change you need to make:

```groovy
android/capacitor-cordova-android-plugins/build.gradle
<<<<<<< HEAD
android {
    compileSdkVersion project.hasProperty('compileSdkVersion') ? rootProject.ext.compileSdkVersion : 30
    defaultConfig {
        minSdkVersion project.hasProperty('minSdkVersion') ? rootProject.ext.minSdkVersion : 21
        targetSdkVersion project.hasProperty('targetSdkVersion') ? rootProject.ext.targetSdkVersion : 30
        versionCode 1
        versionName "1.0"
    }
=======
android {
    namespace 'com.forthexp.doit'
    compileSdkVersion project.hasProperty('compileSdkVersion') ? rootProject.ext.compileSdkVersion : 30
    defaultConfig {
        minSdkVersion project.hasProperty('minSdkVersion') ? rootProject.ext.minSdkVersion : 21
        targetSdkVersion project.hasProperty('targetSdkVersion') ? rootProject.ext.targetSdkVersion : 30
        versionCode 1
        versionName "1.0"
    }
>>>>>>> updated
```

This change adds the `namespace 'com.forthexp.doit'` line to the `android` block. The value `'com.forthexp.doit'` is the same as the `package` attribute in your `AndroidManifest.xml` file.