aider: Added "+" and "-" buttons to the range input that shift the value to the next tick.
# Aider chat conversation:

USER: add + and - buttons to the using font awesome icons that shifts the value of the range to the next tick

ASSISTANT: To add "+" and "-" buttons that shift the value of the range to the next tick, we need to modify the `XpAddAchievement.vue` file.

We will add two `ion-buttons` before and after the `ion-range` component. These buttons will have "+" and "-" icons from Font Awesome. We will also add `@click` event listeners to these buttons that will increase or decrease the value of `achievement.difficulty` respectively.

Here is the *edit block* for the changes:

```html
src/views/App/SideMenu/XpGameMaster/XpAchievements/XpAddAchievement/XpAddAchievement.vue
<<<<<<< HEAD
          <ion-range
            min="1"
            max="13"
            :pin="true"
            mode="ios"
            step="1"
            snaps
            ticks
            color="primary"
            v-model="achievement.difficulty"
            @ionChange="updatePoints"
          >
            <ion-label slot="start">Minimum</ion-label>
            <ion-label slot="end">Maximum </ion-label>
          </ion-range>
=======
          <ion-button @click="achievement.difficulty = Math.max(1, achievement.difficulty - 1)">
            <i class="fas fa-minus"></i>
          </ion-button>
          <ion-range
            min="1"
            max="13"
            :pin="true"
            mode="ios"
            step="1"
            snaps
            ticks
            color="primary"
            v-model="achievement.difficulty"
            @ionChange="updatePoints"
          >
            <ion-label slot="start">Minimum</ion-label>
            <ion-label slot="end">Maximum </ion-label>
          </ion-range>
          <ion-button @click="achievement.difficulty = Math.min(13, achievement.difficulty + 1)">
            <i class="fas fa-plus"></i>
          </ion-button>
>>>>>>> updated
```

Please note that you need to have the Font Awesome library included in your project to use the `fas fa-minus` and `fas fa-plus` classes. If you don't have it yet, you can add it to your project by including the following line in the `<head>` section of your HTML file:

```html
<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
```