aider: Add always-show-pin attribute to ion-range component.
# Aider chat conversation:

USER: i would like to make it so the pin always shows the value of the selected range

ASSISTANT: To make the pin always show the value of the selected range, we need to modify the `ion-range` component in the `XpAddAchievement.vue` file.

We will add the `always-show-pin` attribute to the `ion-range` component. This attribute will make the pin always visible with the current value.

Here is the *edit block* for the changes:

```html
src/views/App/SideMenu/XpGameMaster/XpAchievements/XpAddAchievement/XpAddAchievement.vue
<<<<<<< HEAD
          <ion-range
            min="1"
            max="13"
            :pin="true"
            mode="ios"
            step="1"
            snaps
            ticks
            color="primary"
            v-model="achievement.difficulty"
            @ionChange="updatePoints"
          >
            <ion-label slot="start">Minimum</ion-label>
            <ion-label slot="end">Maximum </ion-label>
          </ion-range>
=======
          <ion-range
            min="1"
            max="13"
            :pin="true"
            always-show-pin
            mode="ios"
            step="1"
            snaps
            ticks
            color="primary"
            v-model="achievement.difficulty"
            @ionChange="updatePoints"
          >
            <ion-label slot="start">Minimum</ion-label>
            <ion-label slot="end">Maximum </ion-label>
          </ion-range>
>>>>>>> updated
```

This change will make the pin always visible with the current value of the range.