aider: Created a new database file `DosDontsDb.ts` to store "dos" and "donts" with methods for setting, deleting, and retrieving them.
# Aider chat conversation:

USER: I need to make another db for The dos and donts. it will essentially hold a do/dont which will have the following fields to it: whats the bonus for - xp, gp, ap, notes. type - do or dont

ASSISTANT: To create a new database for the "dos" and "donts", we will need to create a new TypeScript file. This file will define the structure of the "dos" and "donts" and provide methods for interacting with the data.

Let's create a new file named `DosDontsDb.ts` in the `src/databases/` directory. This file will contain the structure of the "dos" and "donts" and the methods to interact with them.

```typescript
src/databases/DosDontsDb.ts
<<<<<<< HEAD

=======
import { v4 as uuidv4 } from 'uuid'
import DbStorageApi from './DbStorageApi';

import { Drivers, Storage } from "@ionic/storage";

export const dosDontsStorage = new Storage({
  name: "__dosDonts",
  driverOrder: [Drivers.IndexedDB, Drivers.LocalStorage],
});

export interface DosDont {
  id?: any
  type: 'do' | 'dont'
  bonus: {
    xp: number;
    gp: number;
    ap: number;
  }
  notes: string;
}

export class DosDontsDb extends DbStorageApi {
  private createDosDont(): DosDont {
    return {
      id: uuidv4(),
      type: 'do',
      bonus: {
        xp: 0,
        gp: 0,
        ap: 0,
      },
      notes: '',
    }
  }

  public async setDosDont(dosDont: DosDont) {
    const id = dosDont.id ? dosDont.id : uuidv4()
    await this.set(id, {
      ...this.createDosDont(),
      ...dosDont,
      id
    })
  }

  public async deleteDosDont(dosDont: DosDont) {
    this.remove(dosDont.id)
    return await this.getDosDonts()
  }

  public async getDosDontById(id) {
    const dosDonts = await this.getDosDonts();
    const dosDont = dosDonts.find(dosDont => dosDont.id === id);
    return dosDont || this.createDosDont()
  }

  public async getDosDonts() {
    const dosDonts = await this.getAll()
    return dosDonts
  }
}

export default DosDontsDb
>>>>>>> updated
```

This new file `DosDontsDb.ts` will create a new database for the "dos" and "donts". It defines the structure of a "do" or "dont" and provides methods for setting, deleting, and retrieving them.