aider: Added a row about the amount of points to be awarded with three columns of checkboxes for XP, GP, and AP, using their respective icons.
# Aider chat conversation:

USER: add a row about the amount of points to be awareded. make this row 3 cols of checkbox each for Xp, Gp, and Ap.

USER: add a row about the amount of points to be awareded. make this row 3 cols of checkbox each for Xp, Gp, and Ap.

ASSISTANT: To add a row with three columns of checkboxes for XP, GP, and AP,
^C KeyboardInterrupt

USER: add a row about the amount of points to be awareded. make this row 3 cols of checkbox each for Xp, Gp, and Ap. be sure to use their icons on the rows too

ASSISTANT: To add a row with three columns of checkboxes for XP, GP, and AP, we need to modify the `XpAddDoDont.vue` file. We will add a new `ion-row` component after the `ion-list-header` that says "Amount of Points to Award". Inside this row, we will add three `ion-col` components, each containing an `ion-checkbox` and an `ion-label` with the corresponding icon for XP, GP, and AP.

Here is the *edit block* for this change:

```vue
src/views/App/SideMenu/XpGameMaster/XpDoThisNotThat/components/XpAddDoDont.vue
<<<<<<< HEAD
      <ion-list-header> Amount of Points to Award </ion-list-header>
=======
      <ion-list-header> Amount of Points to Award </ion-list-header>
      <ion-row>
        <ion-col>
          <ion-checkbox v-model="form.points.xp" color="success"></ion-checkbox>
          <ion-label>
            <i class="fad fa-hand-holding-seedling"></i> XP
          </ion-label>
        </ion-col>
        <ion-col>
          <ion-checkbox v-model="form.points.gp" color="warning"></ion-checkbox>
          <ion-label>
            <i class="fad fa-hand-holding-usd"></i> GP
          </ion-label>
        </ion-col>
        <ion-col>
          <ion-checkbox v-model="form.points.ap" color="danger"></ion-checkbox>
          <ion-label>
            <i class="fad fa-hand-holding-magic"></i> AP
          </ion-label>
        </ion-col>
      </ion-row>
>>>>>>> updated
```

This change adds a new row with three columns. Each column contains a checkbox and a label with the corresponding icon for XP, GP, and AP. The `v-model` directive is used to create two-way data bindings on the checkbox elements, which means that the state of the checkboxes will be bound to the `form.points.xp`, `form.points.gp`, and `form.points.ap` data properties.