aider: Added selectedTab property to XpGameMaster.ts and bound selected attribute of ion-tab-button elements to selectedTab condition in XpGameMaster.vue.
# Aider chat conversation:

ASSISTANT: Thank you for providing the file. However, the `XpGameMaster.ts` file you provided doesn't contain the necessary code to handle the selected tab state. We need to add a `selectedTab` property to the `setup()` function and return it. This property will hold the current selected tab.

Here is the change needed in `XpGameMaster.ts`:

```typescript
src/views/App/SideMenu/XpGameMaster/XpGameMaster.ts
<<<<<<< HEAD
  setup() {
    // code
    return {
      arrowBack,
      speedometerOutline,
      speedometerSharp,
      trophyOutline,
      trophySharp,
      pawOutline,
      pawSharp
    }
=======
  setup() {
    // code
    const selectedTab = ref('dashboard'); // default selected tab
    return {
      arrowBack,
      speedometerOutline,
      speedometerSharp,
      trophyOutline,
      trophySharp,
      pawOutline,
      pawSharp,
      selectedTab
    }
>>>>>>> updated
```

Then, in `XpGameMaster.vue`, we need to bind the `selected` attribute of each `ion-tab-button` to a condition that checks if the `selectedTab` is equal to the tab name. Here is the change needed in `XpGameMaster.vue`:

```html
src/views/App/SideMenu/XpGameMaster/XpGameMaster.vue
<<<<<<< HEAD
          <ion-tab-button
            tab="abilities"
            href="/game-master/abilities"
          >
=======
          <ion-tab-button
            tab="abilities"
            href="/game-master/abilities"
            :selected="selectedTab === 'abilities'"
          >
>>>>>>> updated
```

Please repeat the above change for all `ion-tab-button` elements, replacing `'abilities'` with the respective tab name.

Please note that this solution assumes that you are using Vue 3 and Ionic Vue. If you are using a different version or library, the solution might be different.